# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API lists for E-Rescues",
        "title": "Swagger UI for E-Rescues",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:19000",
    "basePath": "/api/v1",
    "paths": {
...
        "/quizz/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "output: struct quizzs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizz-manager-apis"
                ],
                "summary": "Get quizzs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/quizz/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "input: quizz model.Quizz =\u003e output: status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizz-manager-apis"
                ],
                "summary": "Create quizzs",
                "parameters": [
                    {
                        "description": "fill quizz",
                        "name": "quizz",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Quizz"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/quizz/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "input: quizz's id =\u003e output: status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizz-manager-apis"
                ],
                "summary": "Delete quizz by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "quizz's id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/quizz/update": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "input: quizz model.Quizz =\u003e output: status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizz-manager-apis"
                ],
                "summary": "Update quizzs",
                "parameters": [
                    {
                        "description": "change quizz",
                        "name": "quizz",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Quizz"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/quizz/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "input: quizz's id =\u003e output: struct quizz",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizz-manager-apis"
                ],
                "summary": "Get quizz by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "quizz's id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Quizz": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "profileId": {
                    "type": "integer"
                },
                "question_group": {
                    "type": "integer"
                },
                "score_anxiety": {
                    "type": "integer"
                },
                "score_depess": {
                    "type": "integer"
                },
                "score_stress": {
                    "description": "Điểm số của bài test",
                    "type": "integer"
                }
            }
        },
        "model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}